+++ lib/ui/view/stage/ml_kit_camera_play_view.dart
@@ -1,20 +1,18 @@
 // 카메라 화면
 import 'dart:async';
 
-import 'package:assets_audio_player/assets_audio_player.dart';
 import 'package:camera/camera.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:google_mlkit_pose_detection/google_mlkit_pose_detection.dart';
 import 'package:pocket_pose/config/app_color.dart';
-import 'package:pocket_pose/config/audio_player/audio_player_util.dart';
 import 'package:pocket_pose/config/ml_kit/custom_pose_painter.dart';
 import 'package:pocket_pose/data/entity/socket_request/send_skeleton_request.dart';
 import 'package:pocket_pose/data/remote/provider/socket_stage_provider_impl.dart';
-import 'package:pocket_pose/data/remote/provider/stage_provider_impl.dart';
 import 'package:pocket_pose/domain/entity/stage_skeleton_pose_landmark.dart';
 import 'package:pocket_pose/main.dart';
 import 'package:pocket_pose/ui/widget/stage/stage_appbar_music_info_widget.dart';
+import 'package:pocket_pose/ui/widget/stage/stage_play_countdown_widget.dart';
 import 'package:provider/provider.dart';
 
 class MlKitCameraPlayView extends StatefulWidget {
@@ -50,22 +48,9 @@ class _MlKitCameraPlayViewState extends State<MlKitCameraPlayView> {
   bool _canProcess = true;
   bool _isBusy = false;
   int _frameNum = 0;
-  // 5초 카운트다운변수
-  bool _countdownVisibility = false;
-  final List<String> _countdownSVG = [
-    'assets/icons/ic_countdown_1.png',
-    'assets/icons/ic_countdown_2.png',
-    'assets/icons/ic_countdown_3.png',
-    'assets/icons/ic_countdown_4.png',
-    'assets/icons/ic_countdown_5.png',
-  ];
-  int _seconds = 5;
-  Timer? _timer;
-  AssetsAudioPlayer? _assetsAudioPlayer;
   // 스켈레톤 추출 변수 선언(google_mlkit_pose_detection 라이브러리)
   final PoseDetector _poseDetector =
       PoseDetector(options: PoseDetectorOptions());
-  late StageProviderImpl _stageProvider;
   late SocketStageProviderImpl _socketStageProvider;
 
   @override
@@ -82,10 +67,8 @@ class _MlKitCameraPlayViewState extends State<MlKitCameraPlayView> {
   @override
   void initState() {
     super.initState();
-    _stageProvider = Provider.of<StageProviderImpl>(context, listen: false);
     _socketStageProvider =
         Provider.of<SocketStageProviderImpl>(context, listen: false);
-    _assetsAudioPlayer = AssetsAudioPlayer();
 
     // 카메라 설정. 기기에서 실행 가능한 카메라, 카메라 방향 설정...
     if (cameras.any(
@@ -107,9 +90,6 @@ class _MlKitCameraPlayViewState extends State<MlKitCameraPlayView> {
       }
     }
 
-    // 중간 입장 처리
-    _onMidEnter();
-
     // 플레이어이면서 카메라 실행 가능하면 포즈 추출 시작
     if (widget.playerNum != -1 && _cameraIndex != -1) {
       _startLiveFeed();
@@ -118,43 +98,11 @@ class _MlKitCameraPlayViewState extends State<MlKitCameraPlayView> {
 
   @override
   void dispose() {
-    _stopTimer();
-    _assetsAudioPlayer = null;
-    _assetsAudioPlayer?.dispose();
     _canProcess = false;
     _poseDetector.close();
     _controller?.dispose;
 
     super.dispose();
-  }
-
-  void _onMidEnter() {
-    (_socketStageProvider.catchMusicData != null)
-        ? AudioPlayerUtil()
-            .setMusicUrl(_socketStageProvider.catchMusicData!.musicUrl)
-        : AudioPlayerUtil().setMusicUrl(_stageProvider.music!.musicUrl);
-
-    // 중간임장인 경우
-    if (widget.midEnterSeconds != -1) {
-      // 중간 입장한 초부터 시작
-      _seconds = widget.midEnterSeconds;
-    } else {
-      // 중간입장 아닐 시 처음부터 시작(0초)
-      _seconds = 0;
-    }
-
-    // 카운트다운
-    if (_seconds < 5) {
-      // 카운트다운 시작 후 노래 재생
-      _seconds = 5 - _seconds;
-      _countdownVisibility = true;
-
-      _startTimer();
-    }
-    // 노래 재생
-    else {
-      AudioPlayerUtil().playSeek(_seconds - 5);
-    }
   }
 
   // 카메라 화면 보여주기 + 화면에서 실시간으로 포즈 추출
@@ -176,7 +124,7 @@ class _MlKitCameraPlayViewState extends State<MlKitCameraPlayView> {
         const StageAppbarMusicInfoWidget(),
         // 추출된 스켈레톤 그리기
         _liveFeedBodyPlay(),
-        buildCountdownWidget(),
+        StagePlayCountdownWidget(midEnterSeconds: widget.midEnterSeconds),
       ],
     );
   }
@@ -346,50 +294,6 @@ class _MlKitCameraPlayViewState extends State<MlKitCameraPlayView> {
           ],
         );
     }
-  }
-
-  Widget buildCountdownWidget() {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      crossAxisAlignment: CrossAxisAlignment.center,
-      children: [
-        Visibility(
-            visible: _countdownVisibility,
-            child: (6 > _seconds) && (_seconds > 0)
-                ? Image.asset(_countdownSVG[_seconds - 1])
-                : Container())
-      ],
-    );
-  }
-
-  void _startTimer() {
-    _assetsAudioPlayer?.open(Audio("assets/audios/sound_play_wait.mp3"));
-
-    _timer = Timer.periodic(const Duration(seconds: 1), (timer) {
-      if (_seconds == 1) {
-        _stopTimer();
-
-        if (mounted) {
-          setState(() {
-            _countdownVisibility = false;
-          });
-        }
-        _seconds = 5;
-        AudioPlayerUtil().play();
-      } else {
-        if (mounted) {
-          _assetsAudioPlayer?.open(Audio("assets/audios/sound_play_wait.mp3"));
-          setState(() {
-            _seconds--;
-          });
-        }
-      }
-    });
-  }
-
-  void _stopTimer() {
-    _timer?.cancel();
-    _timer = null;
   }
 
   // 실시간으로 카메라에서 이미지 받기(비동기적)
